ADDED: the geo dependency update
